fix: missed laf partition as bootloader type
Hint:
to be totally true LAF is NOT a part of the bootloader stack.
LAF (so download mode) is a separate recovery-like partition
and only get started when explicitly triggered by the user or software.
BUT:
the LAF partition must be COMPATIBLE with the
bootloader stack so it is highly involved when flashing the
bootloader. If you do not flash LAF while flashing the rest
you will break the download mode