new: bringing UsU to SALT
This commit is a real impressive one. SALT now supports to
unlock by UsU! Keep in mind that SALT does *NOT* deliver the
UsU file itself to avoid any issues with LG or whoever. That means
you always need the UsU file first before actually you can make
any use of UsU in SALT.

The implementation of UsU in SALT is as much bulletproof as
possible and the way its implemented is like this:

1) check model - if not in list (means not a G4 actually): EXIT
2) let the user choose the flashing method which can be one of:
 - external sdcard (directly on the sdcard - no subfolder)
 - internal storage (/data/media/0/Downloads)
 - on LP and maybe later on MM,N as well we COULD flash
   directly and this has been prepared but not activated yet
3) SALT mounts either the sdcard or the internal storage depending
    on the users choice
4) SALT searches for the UsU file and if it cant be found - EXITs
5) SALT does a test run to abootbak:
- dumping
- store a sha256sum of that
- writing
- dumping again
- store a sha256sum of that
- compare the sha256sum with the predefined UsU hash
- if they do not match - EXIT
6) do the same as in 5) but now with the real *aboot* partition
    (means the real UsU flashing)
7) if all went well - congrats - if not a hint to reflash a KDZ
    is shown