new: adding GPT check for UsU
this enables a new field in SALT displaying the general
GPT compatibility which can be either H815 or H811 or unknown
depending on what GPT is detected.
This makes use of carrier and cust partitions which is a very good
indicator for identifying. Unknown should never happen (if SALT
works in general) but if the whole GPT will be printed in the log