new: overwrite auto detected CR mode
This commit together with the changes in
https://github.com/steadfasterX/lglaf/commit/acfdbfd7f157
implements a new feature allowing to overwrite the detected
Challenge/Response requirement.

lglaf tries to identify the device, when a specific usb id is found
or when a specific protocol version has been detected CR will
be enabled for all requests.
Some firmwares behave different (mainly some stupid OEM ones)
and require authentication even when on an older protocol or
older usb id name.

By default SALT will start like defined in salt.vars -> CRMODE
which is - unless unchanged - "Auto".

This can be overwritten by either changing the variable or within
the Advanced Menu which loops like that (when CRMODE=Auto):

Force CR -> Force no CR -> Set CR Auto -> Force CR -> ....

That means after clicking "Force CR" SALT will be restarted in that
mode and the button changes to "Force no CR" and so on.

fixes #13