fix: ARB detection in rare cases
some bootloader stacks may  the 6th bit set which is not the ARB but
something else (only LG knows). the real ARB setting should happen
in the bit 7 only (maybe extended to 8 if it goes 2 digit).
Now SALT will start reading at bit 7 instead of 6 - this requires much
testing to see if it fits for everyone..